Add basic UDP support
This commit adds basic support for UDP to CAF. It includes a new
datagram servant for the multiplexer, associated handles, message types
and tests. CAF messages sent via UDP are delivered in order but
unreliably.

The receive buffer has a size of 65k which should enable CAF to receive
all regular datagrams. Currently, messages that are bigger than the MTU
will be fragmented by IP.

Message slicing and optional reliability are planned for the future.